+++ janitor/io.py
@@ -7,7 +7,7 @@ from typing import Iterable, Union
 import pandas as pd
 
 from .errors import JanitorError
-from .utils import deprecated_alias, check
+from .utils import deprecated_alias, check, import_message
 
 
 @deprecated_alias(seperate_df="separate_df", filespath="files_path")
@@ -111,3 +111,112 @@ def read_commandline(cmd: str, **kwargs) -> pd.DataFrame:
     else:
         outcome = outcome.stdout
     return pd.read_csv(StringIO(outcome), **kwargs)
+
+
+def xlsx_table(
+    path: str,
+    sheetname: str,
+    table: Union[str, list, tuple] = None,
+) -> Union[pd.DataFrame, dict]:
+    """
+    Returns a DataFrame of values in a table in the Excel file.
+    This applies to an Excel file, where the data range is explicitly
+    specified as a Microsoft Excel table.
+
+    If there is a single table in the sheet, or a string is provided
+    as an argument to the `table` parameter, a pandas DataFrame is returned;
+    if there is more than one table in the sheet,
+    and the `table` argument is `None`, or a list/tuple of names,
+    a dictionary of DataFrames is returned, where the keys of the dictionary
+    are the table names.
+
+    Example:
+
+    ```python
+
+    filename = "excel_table.xlsx"
+
+    # single table
+    jn.xlsx_table(filename, sheetname='Tables', table = 'dCategory')
+
+        CategoryID      Category
+    0           1       Beginner
+    1           2       Advanced
+    2           3      Freestyle
+    3           4    Competition
+    4           5  Long Distance
+
+    # multiple tables:
+    jn.xlsx_table(filename, sheetname = 'Tables', table = ['dCategory', 'dSupplier'])
+
+    {'dCategory':    CategoryID       Category
+                    0           1       Beginner
+                    1           2       Advanced
+                    2           3      Freestyle
+                    3           4    Competition
+                    4           5  Long Distance,
+    'dSupplier':   SupplierID             Supplier        City State                         E-mail
+                    0         GB       Gel Boomerangs     Oakland    CA          gel@gel-boomerang.com
+                    1         CO  Colorado Boomerangs    Gunnison    CO  Pollock@coloradoboomerang.com
+                    2         CC        Channel Craft    Richland    WA                    Dino@CC.com
+                    3         DB        Darnell Booms  Burlington    VT            Darnell@Darnell.com}
+    ```
+
+    :param path: Path to the Excel File.
+    :param sheetname: Name of the sheet from which the tables
+        are to be extracted.
+    :param table: Name of a table, or list of tables in the sheet.
+    :returns: A pandas DataFrame, or a dictionary of DataFrames,
+        if there are multiple arguments for the `table` parameter,
+        or the argument to `table` is `None`.
+    :raises ValueError: If there are no tables in the sheet.
+
+    """  # noqa : E501
+
+    try:
+        from openpyxl import load_workbook
+    except ImportError:
+        import_message(
+            submodule="io",
+            package="openpyxl",
+            conda_channel="conda-forge",
+            pip_install=True,
+        )
+    wb = load_workbook(filename=path, read_only=False, keep_links=False)
+    check("sheetname", sheetname, [str])
+    ws = wb[sheetname]
+
+    contents = ws.tables
+    if not contents:
+        raise ValueError(f"There is no table in `{sheetname}` sheet.")
+
+    if isinstance(table, str):
+        table = [table]
+    if table is not None:
+        check("table", table, [list, tuple])
+        for entry in table:
+            if entry not in contents:
+                raise ValueError(
+                    f"{entry} is not a table in the {sheetname} sheet."
+                )
+        data = (
+            (key, value) for key, value in contents.items() if key in table
+        )
+    else:
+        data = contents.items()
+
+    frame = {}
+    for key, value in data:
+        content = ((cell.value for cell in row) for row in ws[value])
+        if contents[key].headerRowCount == 1:
+            column_names = next(content)
+            content = zip(*content)
+            frame[key] = dict(zip(column_names, content))
+        else:
+            content = zip(*content)
+            frame[key] = {f"C{num}": val for num, val in enumerate(content)}
+
+    if len(frame) == 1:
+        _, frame = frame.popitem()
+        return pd.DataFrame(frame)
+    return {key: pd.DataFrame(value) for key, value in frame.items()}